one-line change in compressColumn() to get and set double data correctly.

+++ basic/src/ncsa/d2k/modules/core/datatype/table/basic/SubsetTableImpl.java
@@ -446,7 +446,7 @@ public class SubsetTableImpl extends MutableTableImpl {
                 break;
 
               case ColumnTypes.DOUBLE:
-                compressedColumn.setFloat(this.getFloat(i, colindex), i);
+                compressedColumn.setDouble(this.getDouble(i, colindex), i);
                 break;
 
               case ColumnTypes.FLOAT:
